update alerts to sweet alerts in sections.

+++ scripts/section.js
@@ -80,7 +80,7 @@ function addSection() {
   })
   .then((response) => response.text())
   .then(responseText => {
-    alert(responseText);
+    Swal.fire("Success", `${responseText}`, "success");
     displaySections(); // updates the table
   }).catch (error => {
     alert('console error.');
@@ -132,7 +132,7 @@ function editSection(button) {
     })
     .then((response) => response.text())
     .then((responseText) => {
-      alert(responseText);
+      Swal.fire("Success", `${responseText}`, "success");
       displaySections();
     });
 
@@ -141,6 +141,28 @@ function editSection(button) {
     // Edit mode
     button.textContent = "Save";
     cells[0].focus();
+
+    const Toast = Swal.mixin({
+      toast: true,
+      position: 'top-end',
+      icon: "info",
+      showConfirmButton: false,
+      timer: 2000,
+      timerProgressBar: true,
+      customClass: {
+        popup: 'colored-toast'
+      },
+      didOpen: (toast) => {
+        toast.addEventListener('mouseenter', Swal.stopTimer);
+        toast.addEventListener('mouseleave', Swal.resumeTimer);
+      }
+    });
+
+    Toast.fire({
+      icon: 'info',
+      title: 'Edit Mode Activated'
+    });
+
     updatedRow.length = 0; // clear the array.
   }
 }
@@ -150,19 +172,28 @@ function deleteSection(button) {
   const row = button.closest('tr');
   const cells = row.querySelectorAll('td');
   const sectionId = cells[0].innerHTML;
-  
-  fetch(sectionEndpoint, {
-    method: 'DELETE',
-    headers: {
-      "Content-type": "application/x-www-form-urlencoded",
-    },
-    body: `section_id=${sectionId}`,
-  })
-  .then((response) => response.text())
-  .then((responseText) => {
-    alert(responseText);
-    displaySections();
-  })
+
+  Swal.fire({
+    title: "Do you want to delete this data?",
+    icon: "question",
+    showCancelButton: true,
+    confirmButtonText: "Delete",
+  }).then((result) => {
+    if (result.isConfirmed) {
+       fetch(sectionEndpoint, {
+        method: 'DELETE',
+        headers: {
+          "Content-type": "application/x-www-form-urlencoded",
+        },
+        body: `section_id=${sectionId}`,
+      })
+      .then((response) => response.text())
+      .then(() => {
+        Swal.fire("Deleted", "", "success");
+        displaySections();
+      })
+    }
+  });
 }
 
 function searchSection() {
